fix: remove problematic route group re-export pattern
Move landing page content directly to [locale]/page.tsx and remove
the (unauth) route group. The re-export pattern was causing Next.js
build failures on Vercel due to client reference manifest conflicts.
