--- src/app/[locale]/(unauth)/page.tsx
@@ -1,44 +1,0 @@
-import { getTranslations, unstable_setRequestLocale } from 'next-intl/server';
-
-import { CTA } from '@/templates/CTA';
-import { FAQ } from '@/templates/FAQ';
-import { Features } from '@/templates/Features';
-import { Footer } from '@/templates/Footer';
-import { Hero } from '@/templates/Hero';
-import { Navbar } from '@/templates/Navbar';
-
-// Force dynamic rendering to avoid RSC serialization issues during build
-export const dynamic = 'force-dynamic';
-
-export async function generateMetadata(props: { params: Promise<{ locale: string }> }) {
-  const { locale } = await props.params;
-  const t = await getTranslations({
-    locale,
-    namespace: 'Index',
-  });
-
-  return {
-    title: t('meta_title'),
-    description: t('meta_description'),
-  };
-}
-
-const IndexPage = async (props: { params: Promise<{ locale: string }> }) => {
-  const { locale } = await props.params;
-  unstable_setRequestLocale(locale);
-
-  // AC #4: Auth state is now checked client-side for better performance
-  // Landing page is statically generated for fast load times and SEO
-  return (
-    <>
-      <Navbar />
-      <Hero />
-      <Features />
-      <FAQ />
-      <CTA />
-      <Footer />
-    </>
-  );
-};
-
-export default IndexPage;
